edit params; no user id passed needed

+++ src/services/clubs.service.ts
@@ -77,22 +77,22 @@ export class ClubService extends BaseService {
       // Get current user
       const currentUserID = await this.getCurrentUserId();
 
-      // Get user info
-      const { data: user, error: userError } = await this.supabase
-        .from('users')
-        .select('id')
-        .eq('id', currentUserID)
-        .single();
+      // // Get user info
+      // const { data: user, error: userError } = await this.supabase
+      //   .from('users')
+      //   .select('name')
+      //   .eq('id', currentUserID)
+      //   .single();
 
-      // if user not found, throuw error
-      if (userError) throw new Error('User not found');
+      // // if user not found, throuw error
+      // if (userError) throw new Error('User not found');
 
       // Create club
       const { data: club, error } = await this.supabase
         .from('clubs')
         .insert({
           ...data,
-          created_by: user.id,
+          created_by: currentUserID,
           experience_level: 1,
           experience_points: 0,
         })
@@ -103,7 +103,7 @@ export class ClubService extends BaseService {
 
       // Auto-add creator as member
       const { error: membershipError } = await this.supabase.from('club_memberships').insert({
-        user_id: user.id,
+        user_id: currentUserID,
         club_id: club.id,
         role: 'admin',
       });
@@ -158,13 +158,9 @@ export class ClubService extends BaseService {
   }
 
   // Update club mon picture (avatar_url)
-  async updateClubSprite(userId: number, clubId: number, spriteUrl: string): Promise<Club> {
+  async updateClubSprite(clubId: number, spriteUrl: string): Promise<Club> {
     try {
       const currentUserID = await this.getCurrentUserId();
-
-      if (currentUserID !== userId) {
-        throw new Error('Only club admins can edit club sprite!!');
-      }
 
       // Check membership join table for user's role
       const { data: membership, error: membershipError } = await this.supabase
@@ -198,13 +194,9 @@ export class ClubService extends BaseService {
   }
 
   // Update club mon picture (avatar_url)
-  async updateClubLogo(userId: number, clubId: number, logoUrl: string): Promise<Club> {
+  async updateClubLogo(clubId: number, logoUrl: string): Promise<Club> {
     try {
       const currentUserID = await this.getCurrentUserId();
-
-      if (currentUserID !== userId) {
-        throw new Error('Only club admins can edit club sprite!!');
-      }
 
       // Check membership join table for user's role
       const { data: membership, error: membershipError } = await this.supabase
@@ -240,37 +232,37 @@ export class ClubService extends BaseService {
   // Delete club; only admins can do so
   async deleteClub(clubId: number): Promise<void> {
     try {
-      // const currentClubID = await this.getCurrentUserId();
+      const currentClubID = await this.getCurrentUserId();
 
-      // // Get user profile for checking if user has permissions
-      // const { data: user, error: userError } = await this.supabase
-      //   .from('users')
-      //   .select('id, user_type')
-      //   .eq('id', currentClubID)
-      //   .single();
+      // Get user profile for checking if user has permissions
+      const { data: user, error: userError } = await this.supabase
+        .from('users')
+        .select('id, user_type')
+        .eq('id', currentClubID)
+        .single();
 
-      // // Failed to get user
-      // if (userError || !user) {
-      //   throw new Error('User profile not found for deleting club');
-      // }
+      // Failed to get user
+      if (userError || !user) {
+        throw new Error('User profile not found for deleting club');
+      }
 
-      // // Get club info
-      // const { data: existingClub, error: clubError } = await this.supabase
-      //   .from('clubs')
-      //   .select('created_by')
-      //   .eq('id', clubId)
-      //   .single();
+      // Get club info
+      const { data: existingClub, error: clubError } = await this.supabase
+        .from('clubs')
+        .select('created_by')
+        .eq('id', clubId)
+        .single();
 
-      // if (clubError) throw clubError;
+      if (clubError) throw clubError;
 
-      // // Check if user has permission to update this club (admin)
-      // if (
-      //   existingClub != null &&
-      //   existingClub.created_by !== user.id &&
-      //   user.user_type !== 'admin'
-      // ) {
-      //   throw new Error('You do not have permission to delete this club');
-      // }
+      // Check if user has permission to update this club (admin)
+      if (
+        existingClub != null &&
+        existingClub.created_by !== user.id &&
+        user.user_type !== 'admin'
+      ) {
+        throw new Error('You do not have permission to delete this club');
+      }
 
       const { error: deleteError } = await this.supabase.from('clubs').delete().eq('id', clubId);
 
@@ -281,34 +273,6 @@ export class ClubService extends BaseService {
       this.handleError(error, 'ClubService.deleteClub');
     }
   }
-
-  // Get members of club
-  // async getClubMembers(clubId: number, limit = 10, offset = 0): Promise<ClubMember[]> {
-  //   try {
-  //     const { data, error } = await this.supabase
-  //       .from('club_memberships')
-  //       .select(
-  //         `
-  //         *,
-  //         users:user_id (
-  //           id,
-  //           first_name,
-  //           last_name,
-  //           avatar_url
-  //         )
-  //       `,
-  //       )
-  //       .eq('club_id', clubId)
-  //       .range(offset, offset + limit - 1);
-
-  //     if (error) throw error;
-
-  //     return data;
-  //   } catch (error: any) {
-  //     this.handleError(error, 'ClubService.getClubMembers');
-  //     return [];
-  //   }
-  // }
 }
 
 // Export singleton instance for reusing across calls
